refactor: move presentation logic of User and Article, Profile models into decorators
