+++ app/decorators/profile_decorator.rb
@@ -1,0 +1,13 @@
+module ProfileDecorator
+  def age
+    return '不明' unless birthday.present?
+    years = Time.zone.now.year - birthday.year
+    days = Time.zone.now.yday - birthday.yday
+
+    if days < 0
+      "#{years -1}歳"
+    else
+      "#{years}歳"
+    end
+  end
+end
