+++ app/models/profile.rb
@@ -3,15 +3,4 @@ class Profile < ApplicationRecord
   belongs_to :user
   has_one_attached :avatar
 
-  def age
-    return '不明' unless birthday.present?
-    years = Time.zone.now.year - birthday.year
-    days = Time.zone.now.yday - birthday.yday
-
-    if days < 0
-      "#{years -1}歳"
-    else
-      "#{years}歳"
-    end
-  end
 end
